traffsoft-backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy[asyncio]
greenlet
asyncpg
psycopg2-binary
aiosqlite

python-dotenv
pydantic
pydantic-settings
email-validator

python-jose[cryptography]
passlib[bcrypt]

alembic

httpx
